done the testing API running as expected

+++ DeleteOper_164/DeleteBranchDetails/src/main/java/com/amdocs/controller/BranchDetailsController.java
@@ -14,7 +14,7 @@ public class BranchDetailsController {
 	private BranchDetailsService branchDetailsService;
 	
 	@DeleteMapping("/delete/{id}")
-	public void deleteBranchDetailsById(@PathVariable("id") Long id) {
+	public void deleteBranchDetailsById(@PathVariable("id") Integer id) {
 		branchDetailsService.deleteBranchDetails(id);
 	}
