+++ DeleteOper_164/DeleteBranchDetails/src/main/java/com/amdocs/service/BranchDetailsService.java
@@ -1,9 +1,9 @@
 package com.amdocs.service;
 
-import com.amdocs.model.BranchDetails;
+
 
 public interface BranchDetailsService {
 	
-	public void deleteBranchDetails(Long id);
+	public void deleteBranchDetails(Integer id);
 
 }
