+++ DeleteOper_164/DeleteBranchDetails/src/main/java/com/amdocs/serviceimpl/BranchDetailsServiceImpl.java
@@ -3,7 +3,7 @@ package com.amdocs.serviceimpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.amdocs.model.BranchDetails;
+
 import com.amdocs.repository.BranchDetailsRepository;
 import com.amdocs.service.BranchDetailsService;
 
@@ -14,7 +14,7 @@ public class BranchDetailsServiceImpl implements BranchDetailsService{
 	private BranchDetailsRepository branchDetailsRepostiory;
 	
 	@Override
-	public void deleteBranchDetails(Long id) {
+	public void deleteBranchDetails(Integer id) {
 		branchDetailsRepostiory.deleteById(id);
 	}
